backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.29
asyncpg==0.29.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.2.1
line-bot-sdk==3.10.1
requests==2.31.0
python-multipart==0.0.9
jinja2==3.1.3
uuid==1.30
numpy==1.26.4
faker==24.4.0
google-generativeai==0.4.1
yt-dlp
